better error formatting setup
