+++ packages/gatsby-core-utils/src/errors.ts
@@ -1,102 +1,12 @@
-import PrettyError from "pretty-error"
-import stackTrace from "stack-trace"
-import { IStructuredStackFrame } from "gatsby-reporter"
-import { prepareStackTrace, ErrorWithCodeFrame } from "./prepare-stack-trace"
-import { isNodeInternalModulePath } from "./path"
+import { prepareStackTrace } from "./prepare-stack-trace"
 
-const packagesToSkip = [`core-js`, `bluebird`, `regenerator-runtime`, `graphql`]
+export class ErrorWithCodeFrame extends Error {
+  codeFrame: string
 
-const packagesToSkipTest = new RegExp(
-  `node_modules[\\/](${packagesToSkip.join(`|`)})`
-)
-
-// TO-DO: move this this out of this file (and probably delete this file completely)
-// it's here because it re-implements similar thing as `pretty-error` already does
-export const sanitizeStructuredStackTrace = (
-  stack: stackTrace.StackFrame[]
-): IStructuredStackFrame[] => {
-  // first filter out not useful call sites
-  stack = stack.filter(callSite => {
-    if (!callSite.getFileName()) {
-      return false
-    }
-
-    if (packagesToSkipTest.test(callSite.getFileName())) {
-      return false
-    }
-
-    if (callSite.getFileName().includes(`asyncToGenerator.js`)) {
-      return false
-    }
-
-    if (isNodeInternalModulePath(callSite.getFileName())) {
-      return false
-    }
-
-    return true
-  })
-
-  // then sanitize individual call site objects to make sure we don't
-  // emit objects with extra fields that won't be handled by consumers
-  return stack.map(callSite => {
-    return {
-      fileName: callSite.getFileName(),
-      functionName: callSite.getFunctionName(),
-      columnNumber: callSite.getColumnNumber(),
-      lineNumber: callSite.getLineNumber(),
-    }
-  })
-}
-
-type PrettyRenderError = Error | ErrorWithCodeFrame
-
-export function getErrorFormatter(): PrettyError {
-  const prettyError = new PrettyError()
-  const baseRender = prettyError.render
-
-  prettyError.skipNodeFiles()
-  prettyError.skipPackage(
-    `regenerator-runtime`,
-    `graphql`,
-    `core-js`
-    // `static-site-generator-webpack-plugin`,
-    // `tapable`, // webpack
-  )
-
-  // @ts-ignore the type defs in prettyError are wrong
-  prettyError.skip((traceLine: any) => {
-    if (traceLine && traceLine.file === `asyncToGenerator.js`) return true
-    return false
-  })
-
-  prettyError.appendStyle({
-    "pretty-error": {
-      marginTop: 1,
-    },
-    "pretty-error > header": {
-      background: `red`,
-    },
-    "pretty-error > header > colon": {
-      color: `white`,
-    },
-  })
-
-  if (process.env.FORCE_COLOR === `0`) {
-    prettyError.withoutColors()
+  constructor(...args) {
+    super(...args)
+    this.codeFrame = ``
   }
-
-  prettyError.render = (
-    err: PrettyRenderError | PrettyRenderError[]
-  ): string => {
-    if (Array.isArray(err)) {
-      return err.map(e => prettyError.render(e)).join(`\n`)
-    }
-
-    let rendered = baseRender.call(prettyError, err)
-    if (`codeFrame` in err) rendered = `\n${err.codeFrame}\n${rendered}`
-    return rendered
-  }
-  return prettyError
 }
 
 /**
