+++ packages/gatsby-core-utils/src/index.ts
@@ -6,8 +6,4 @@ export { getCIName, isCI } from "./ci"
 export { createRequireFromPath } from "./create-require-from-path"
 export { getConfigStore } from "./get-config-store"
 export { getGatsbyVersion } from "./get-gatsby-version"
-export {
-  sanitizeStructuredStackTrace,
-  getErrorFormatter,
-  createErrorFromString,
-} from "./errors"
+export { ErrorWithCodeFrame, createErrorFromString } from "./errors"
